[update] ResultView 우승자 출력 및 자동차 이름 출력
